im-util: Make previously-#[doc(hidden)] types and traits unnamable
