+++ eyeball-im-util/src/vector.rs
@@ -2,12 +2,12 @@
 
 mod filter;
 mod limit;
-
-use std::collections::VecDeque;
+mod ops;
 
 use eyeball_im::VectorDiff;
 use futures_core::Stream;
 
+use self::ops::{VectorDiffContainerFamily, VectorDiffContainerOps};
 pub use self::{
     filter::{Filter, FilterMap},
     limit::{EmptyLimitStream, Limit},
@@ -43,99 +43,3 @@ type VectorDiffContainerStreamFamily<S> =
 /// mapped to the given type `U`, from a stream of [`VectorDiffContainer`]s.
 pub type VectorDiffContainerStreamMappedItem<S, U> =
     <VectorDiffContainerStreamFamily<S> as VectorDiffContainerFamily>::Member<U>;
-
-#[doc(hidden)]
-pub trait VectorDiffContainerOps<T> {
-    type Family: VectorDiffContainerFamily;
-
-    fn from_item(vector_diff: VectorDiff<T>) -> Self;
-
-    fn pick_item(vector_diffs: &mut VecDeque<Self>) -> Option<Self>
-    where
-        Self: Sized;
-
-    fn for_each(self, f: impl FnMut(VectorDiff<T>));
-
-    fn filter_map<U>(
-        self,
-        f: impl FnMut(VectorDiff<T>) -> Option<VectorDiff<U>>,
-    ) -> Option<<Self::Family as VectorDiffContainerFamily>::Member<U>>;
-}
-
-impl<T> VectorDiffContainerOps<T> for VectorDiff<T> {
-    type Family = VectorDiffFamily;
-
-    fn from_item(vector_diff: VectorDiff<T>) -> Self {
-        vector_diff
-    }
-
-    fn pick_item(vector_diffs: &mut VecDeque<Self>) -> Option<Self>
-    where
-        Self: Sized,
-    {
-        vector_diffs.pop_front()
-    }
-
-    fn for_each(self, mut f: impl FnMut(VectorDiff<T>)) {
-        f(self);
-    }
-
-    fn filter_map<U>(
-        self,
-        mut f: impl FnMut(VectorDiff<T>) -> Option<VectorDiff<U>>,
-    ) -> Option<<Self::Family as VectorDiffContainerFamily>::Member<U>> {
-        f(self)
-    }
-}
-
-impl<T> VectorDiffContainerOps<T> for Vec<VectorDiff<T>> {
-    type Family = VecVectorDiffFamily;
-
-    fn from_item(vector_diff: VectorDiff<T>) -> Self {
-        vec![vector_diff]
-    }
-
-    fn pick_item(vector_diffs: &mut VecDeque<Self>) -> Option<Self>
-    where
-        Self: Sized,
-    {
-        Some(vector_diffs.drain(..).flatten().collect())
-    }
-
-    fn for_each(self, f: impl FnMut(VectorDiff<T>)) {
-        self.into_iter().for_each(f);
-    }
-
-    fn filter_map<U>(
-        self,
-        f: impl FnMut(VectorDiff<T>) -> Option<VectorDiff<U>>,
-    ) -> Option<<Self::Family as VectorDiffContainerFamily>::Member<U>> {
-        let res: Vec<_> = self.into_iter().filter_map(f).collect();
-        if res.is_empty() {
-            None
-        } else {
-            Some(res)
-        }
-    }
-}
-
-#[doc(hidden)]
-pub trait VectorDiffContainerFamily {
-    type Member<T>: VectorDiffContainerOps<T, Family = Self>;
-}
-
-#[doc(hidden)]
-#[derive(Debug)]
-pub enum VectorDiffFamily {}
-
-impl VectorDiffContainerFamily for VectorDiffFamily {
-    type Member<T> = VectorDiff<T>;
-}
-
-#[doc(hidden)]
-#[derive(Debug)]
-pub enum VecVectorDiffFamily {}
-
-impl VectorDiffContainerFamily for VecVectorDiffFamily {
-    type Member<T> = Vec<VectorDiff<T>>;
-}
